+++ src/float/big.rs
@@ -7195,9 +7195,17 @@ impl Float {
     /// smaller possible results, many bits will be zero, and not all
     /// the precision will be used.
     ///
-    /// `Assign<Src> for Result<Float, Float>` and
-    /// `Assign<Src> for Result<&mut Float, &mut Float>` are
-    /// implemented with the returned object as `Src`.
+    /// There is a corner case where the generated random number is
+    /// converted to NaN: if the precision is very large, the
+    /// generated random number could have an exponent less than the
+    /// allowed minimum exponent, and NaN is used to indicate this.
+    /// For this to occur in practice, the minimum exponent has to be
+    /// set to have a very small magnitude using the low-level MPFR
+    /// interface, or the random number generator has to be designed
+    /// specifically to trigger this case.
+    ///
+    /// `Assign<Src> for Float` is implemented with the returned
+    /// object as `Src`.
     ///
     /// # Examples
     ///
@@ -7206,23 +7214,10 @@ impl Float {
     /// use rug::rand::RandState;
     /// let mut rand = RandState::new();
     /// let mut f = Float::new(2);
-    /// {
-    ///     let mut res = Ok(&mut f);
-    ///     res.assign(Float::random_bits(&mut rand));
-    ///     assert!(res.is_ok());
-    /// }
+    /// f.assign(Float::random_bits(&mut rand));
     /// assert!(f == 0.0 || f == 0.25 || f == 0.5 || f == 0.75);
     /// println!("0.0 ≤ {} < 1.0", f);
     /// ```
-    ///
-    /// # Errors
-    ///
-    /// In all the normal cases, the result will be exact. However, if
-    /// the precision is very large, and the generated random number
-    /// is very small, this may require an exponent smaller than
-    /// [`float::exp_min()`](float/fn.exp_min.html); in this case, the
-    /// number is set to NaN and an error is returned. This would most
-    /// likely be a programming error.
     #[inline]
     pub fn random_bits<'a, 'b: 'a>(
         rng: &'a mut RandState<'b>,
@@ -7232,52 +7227,21 @@ impl Float {
 
     #[cfg(feature = "rand")]
     /// Generates a random number in the range 0 ≤ *x* < 1.
-    ///
-    /// This method is deprecated. The code
-    ///
-    /// ```rust
-    /// use rug::Float;
-    /// let mut f: Float;
-    /// // ...
-    /// # f = Float::new(53);
-    /// # let mut rand = ::rug::rand::RandState::new();
-    /// # let rng = &mut rand;
-    /// # #[allow(deprecated)]
-    /// match f.assign_random_bits(rng) {
-    ///     Ok(()) => { /* ok */ }
-    ///     Err(()) => { /* error */ }
-    /// }
-    /// ```
-    ///
-    /// can be replaced with
-    ///
-    /// ```rust
-    /// use rug::{Assign, Float};
-    /// let mut f: Float;
-    /// // ...
-    /// # f = Float::new(53);
-    /// # let mut rand = ::rug::rand::RandState::new();
-    /// # let rng = &mut rand;
-    /// let mut result = Ok(&mut f);
-    /// result.assign(Float::random_bits(rng));
-    /// match result {
-    ///     Ok(_) => { /* ok */ }
-    ///     Err(_) => { /* error */ }
-    /// };
-    /// ```
     #[deprecated(since = "0.9.2",
-                 note = "use `random_bits` instead; see documentation for an \
-                         example replacement.")]
+                 note = "use `random_bits` instead; \
+                         `f.assign_random_bits(rng)` can be replaced with \
+                         `f.assign(Float::random_bits(rng))`, and testing the \
+                         result can be replaced with testing for NaN.")]
     #[inline]
     pub fn assign_random_bits(
         &mut self,
         rng: &mut RandState,
     ) -> Result<(), ()> {
-        let mut r = Ok(self);
-        r.assign(Float::random_bits(rng));
-        match r {
-            Ok(_) => Ok(()),
-            Err(_) => Err(()),
+        self.assign(Float::random_bits(rng));
+        if self.is_nan() {
+            Err(())
+        } else {
+            Ok(())
         }
     }
 
@@ -7704,18 +7668,12 @@ pub struct RandomBits<'a, 'b: 'a> {
 }
 
 #[cfg(feature = "rand")]
-impl<'a, 'b: 'a, 'c> Assign<RandomBits<'a, 'b>>
-    for Result<&'c mut Float, &'c mut Float>
-{
+impl<'a, 'b: 'a, 'c> Assign<RandomBits<'a, 'b>> for Float {
     #[inline]
     fn assign(&mut self, src: RandomBits<'a, 'b>) {
-        let err = match *self {
-            Ok(ref mut dst) | Err(ref mut dst) => unsafe {
-                mpfr::urandomb(dst.inner_mut(), src.rng.inner_mut())
-            },
-        };
-        if (err != 0) != self.is_err() {
-            misc::result_swap(self)
+        unsafe {
+            let err = mpfr::urandomb(self.inner_mut(), src.rng.inner_mut());
+            assert_eq!(self.is_nan(), err != 0);
         }
     }
 }